chore: Remove env:create, consolidate into env:set
env:set now handles all env:create functionality including
per-environment flags (--staging-value, --dev-value, --staging-inherit,
--dev-inherit, --dev-inherit-staging) for global variables.
