+++ README.md
@@ -33,7 +33,6 @@ solidactions deploy <project-name> <path>
 | `logs:build <project>` | View build/deployment logs |
 | `env:set <key> <value>` | Set a global variable (create or update) |
 | `env:set <project> <key> <value>` | Set a project variable (create or update) |
-| `env:create <key> <value>` | Create a global environment variable |
 | `env:list [project]` | List environment variables |
 | `env:delete <key>` | Delete an environment variable |
 | `env:map <project> <key> <global-key>` | Map a global variable to a project |
